Only call rm if there are files to delete

+++ lib/inventory-rake-1.0/tasks/clean.rb
@@ -50,7 +50,7 @@ class Inventory::Rake::Tasks::Clean
   # Defines the task {#name} that’ll delete {#files}.
   def define
     task name do
-      rm files, :force => true
+      rm files, :force => true unless files.empty?
     end
   end
